requirements.txt
pytest==8.0.2
pytest-html==4.1.1
pytest-bdd==6.1.1
pytest-xdist==3.5.0
pytest-timeout==2.3.1
pytest-rerunfailures==12.0
python-dotenv==1.0.0
pytest-metadata==3.1.0
allure-pytest==2.13.2
playwright>=1.40.0
psutil>=5.9.0
